[fix] change the int type by the float type in the function

+++ src/print_gens.py
@@ -1,7 +1,7 @@
-def print_gen(x: int, y: int) -> None:
+def print_gen(x: float, y: float) -> None:
     print(x, y)
 
 
-def print_gens(xs: list[int], ys: list[int]) -> None:
+def print_gens(xs: list[float], ys: list[float]) -> None:
     for x, y in zip(xs, ys):
         print_gen(x, y)
